requirements.txt
--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

# libs
gdata==1.3.3
Pygments==1.2.2
python-openid==2.2.4
python-twitter==0.6
python-yadis==1.1.0
pytz==2009u
simplejson==2.0.9
textile==2.1.3
vobject==0.8.1c
Markdown==2.0.3
ybrowserauth==1.2
geopy==0.93dev-r0
diff-match-patch==20090804
creole==1.0.1

# externals apps
django-atomformat==0.1.1dev
django-ajax-validation==0.1.3
django-announcements==0.1.0
django-avatar==1.0.3
django-dbtemplates==0.5.8
django-notification==0.1.4
django-pagination==1.0.5.1
django-robots==0.6.1
django-messages==0.4.2
django-threadedcomments==0.5.2
django-extensions==0.4.1
django-uni-form==0.7.0
django-bookmarks==0.1.0
django-gravatar==0.1.0
django-mailer==0.1.0
django-friends==0.1.4
django-locations==0.1.1
django-oembed==0.1.1
django-swaps==0.5.3
django-flag==0.1.1
django-db-log==1.0.1
django-attachments==0.3
django-markup==0.3
django-microblogging==0.2.dev1
django-wikiapp==0.2.0
sorl==3.1
django-generic-flatblocks==0.3.0
django-frontendadmin==0.4
django-filter==0.5.3.dev1
django-photologue==2.3
django-tagging==0.3.1
django-staticfiles==0.2.0
django-debug-toolbar==0.8.3
django-tagging-ext==0.2.0

# packages pinax creates for release:
django-voting==0.1-pre
django-sorting==0.1
template-utils==0.4p2

# pinax dev releases
# careful with these, default pinax uses '>' which can break
django-groups==0.1.dev6
biblion==0.1.dev6
django-email-confirmation==0.1.4
django-openid==0.2.3.dev1
django-timezones==0.2.dev1

# pinax project libs
python-dateutil==1.4.1
docutils==0.5
smartypants==1.6.0.3


# added libs
boto==1.9b

# added apps
django-autoslug==1.3.9
django-build==0.2
django-chronograph==svn-r21-git.2

